requirements.txt
openai
anthropic
gitpython
pydantic
rich
python-dotenv
replicate
pyyaml
pytest